chore: refactor variable methods into new file
